Added TopToolbarField component
